Move simulator -> cartpole.simulator.pydrake, simulator_v2 -> cartpole.simulator.pytorch
