+++ cartpole/device/wire_interface.py
@@ -8,7 +8,7 @@ import varint
 import threading
 from typing import Union, Type, Any
 
-import device.controller_pb2 as proto
+import cartpole.device.controller_pb2 as proto
 from cartpole.common.interface import Error, Config, State
 import os
