+++ cartpole/simulator/pydrake/__init__.py
@@ -1,0 +1,1 @@
+import cartpole.simulator.pydrake.simulator import CartPoleSimulator
