+++ pyproject.toml
@@ -14,6 +14,7 @@ torch = "^1.13.0"
 [tool.poetry.group.dev.dependencies]
 jupyter = "^1.0.0"
 pytest = "^7.2.0"
+matplotlib = "^3.6.2"
 
 [build-system]
 requires = ["poetry-core"]
