--- simulator_v2/pyproject.toml
@@ -1,24 +1,0 @@
-[tool.poetry]
-name = "cartpole-torch"
-version = "0.1.0"
-description = ""
-authors = ["Nikolay Chechulin <nchechulin@protonmail.com>"]
-
-[tool.poetry.dependencies]
-python = "^3.10"
-jupyter = "^1.0.0"
-torch = "^1.12.1"
-matplotlib = "^3.5.3"
-
-[tool.poetry.dev-dependencies]
-isort = "^5.10.1"
-black = {extras = ["ipynb"], version = "^22.6.0"}
-bandit = "^1.7.4"
-flake8 = "^5.0.4"
-mypy = "^0.971"
-pylint = "^2.14.5"
-pre-commit = "^2.20.0"
-
-[build-system]
-requires = ["poetry-core>=1.0.0"]
-build-backend = "poetry.core.masonry.api"
